Move main test files up for shorter paths
