+++ tests/js-only.test.ts
@@ -1,9 +1,9 @@
 import assert from 'node:assert/strict';
 import test from 'node:test';
-import { main } from '../../src/index.js';
-import { resolve, join } from '../../src/util/path.js';
-import baseArguments from '../helpers/baseArguments.js';
-import baseCounters from '../helpers/baseCounters.js';
+import { main } from '../src/index.js';
+import { resolve, join } from '../src/util/path.js';
+import baseArguments from './helpers/baseArguments.js';
+import baseCounters from './helpers/baseCounters.js';
 
 test('Find unused files and exports with JS entry file', async () => {
   const cwd = resolve('tests/fixtures/js-only');
